requirements.txt
humanize
numcodecs
scikit-allel
scipy
zarr